+++ packages/dsi-logo-maker/functions/logoAnimatedCanvas/index.js
@@ -159,7 +159,7 @@ export const presets = {
 };
 
 export const settings = {
-  engine: require("@designsystemsinternational/mechanic-engine-canvas"),
+  engine: require("@mechanic-design/engine-canvas"),
   animated: true,
   usesRandom: true,
 };