+++ packages/dsi-logo-maker/functions/logoAnimatedSVG/index.js
@@ -173,7 +173,7 @@ export const presets = {
 };
 
 export const settings = {
-  engine: require("@designsystemsinternational/mechanic-engine-react"),
+  engine: require("@mechanic-design/engine-react"),
   animated: true,
   usesRandom: true,
 };