+++ packages/dsi-logo-maker/functions/textAnimatedCanvas/index.js
@@ -143,6 +143,6 @@ export const presets = {
 };
 
 export const settings = {
-  engine: require("@designsystemsinternational/mechanic-engine-canvas"),
+  engine: require("@mechanic-design/engine-canvas"),
   animated: true,
 };