+++ packages/dsi-logo-maker/functions/textCanvas/index.js
@@ -99,5 +99,5 @@ export const presets = {
 };
 
 export const settings = {
-  engine: require("@designsystemsinternational/mechanic-engine-canvas"),
+  engine: require("@mechanic-design/engine-canvas"),
 };