+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2021-08-24
+
+### Changed
+
+- Renamed packages and scope. Now all packages live under `@mechanic-design`.
+
 ## [0.4.0] - 2021-08-24
 
 ### Added
 
 - Create CLI project templates and examples
 
-### Change
+### Changed
 
 - Improved project creation CLI experience.
 - Changed single bundle for design functions approach to one isolated bundle per design function.
@@ -59,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First logged release
 
-[unreleased]: https://github.com/designsystemsinternational/mechanic/compare/v0.4.0...main
+[unreleased]: https://github.com/designsystemsinternational/mechanic/compare/v0.4.1...main
+[0.4.1]: https://github.com/designsystemsinternational/mechanic/releases/tag/v0.4.1
 [0.4.0]: https://github.com/designsystemsinternational/mechanic/releases/tag/v0.4.0
 [0.3.0]: https://github.com/designsystemsinternational/mechanic/releases/tag/v0.3.0
 [0.2.11]: https://github.com/designsystemsinternational/mechanic/releases/tag/v0.2.11