+++ packages/dsi-logo-maker/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.4.1 - 2021-08-24
+
+### Changed
+
+- Renamed packages and scope to `@mechanic-design/dsi-logo-maker` and changed to private.
+
 ## 0.4.0 - 2021-08-24
 
 ### Changed